Ignoring records marked as derived records (see last commit) in analysis by default. The user can choose to include these records using the --include_derived_mutations flag to run_analysis.py.
Note:
We say 'derived' rather than 'reversed' as these records are a special case. Here, The mutation protein A -> protein B with DDG x is reversed to be B->A and DDG -x. The general case is that we have DDG values A->B and A->C with DDG values x and y respectively with x = DG(B) - DG(A) and y = DG(C) - DG(A). The mutation B->C with DDG value z = DG(C) - DG(B) can then be derived as y - x and similarly for C->B. These more general cases occur in the SKEMPI dataset for binding affinity. If we represent a derived DDG as a pair of values (y, x) where the derived value is y - x then the reverse mutations are the special case where y = 0 and the original mutations are the special case where x = 0. Indeed, we store DDG values in the database as pairs of values in this manner to avoid redundancy/errors that could occur by storing the experimental values more than once.

Also, while these double-counted values are not necessarily good for benchmarking, it is worth noting that the mutant structures have been determined which may be useful for other purposes.