Starting to add more detailed analysis. We are now starting to generate multiple plots to analyze single runs.
* Feedback analysis *
We now run a simple calculation which aims to choose a scalar value with which to scale the predicted values such that fraction correct is improved (which should also improve the MAE results for metrics with good correlation). On my test set (Kellogg dataset, Rosetta r58124, beta_july15 score function), the fraction correct improved from 0.642 to 0.724 while the MAE improved from 2.816 to 0.834, using a scaling factor of 1/2.908

* Feedback analysis graphs *
Two graphs are created during the feedback analysis:
 - optimum_fraction_correct_at_(experimental cutoff)_kcal_mol plots the fraction correct metric over varying cutoffs defining predicted DDG neutrality
 - optimum_fraction_correct_at_varying_kcal_mol plots how the approximate-best predicted DDG neutrality cutoff varies over different experimental neutrality cutoffs

* Breakdown scatterplots *
I started work creating variants of the exp. vs. pred. scatterplots depending on different criteria. This commit includes:
 - breakdown by exposure using a user-defined burial cutoff
 - breakdown by whether or not a case includes glycine or proline as a wildtype or mutant residue

* Adding more data per dataset case *
 - fields which could be derived but are included for convenience: absolute error, stability classification
 - wt residue, mutant residue
 - wt burial class: one of B/E/None rather than the float value in WildTypeDSSPExposure
 - residue charge change (None, Change, Polar/Charged, Hydrophobic/Non-polar), residue volume change (None, XX, SL, LS)
 - PDB file resolution, PDB resolution bin (currently fixed values: [NA, <1.5, 1.5-<2.0, 2.0-<2.5, >=2.5])
 - PDB chain length

* User interface *
The user can specify the value of burial_cutoff (the value below which residues are considered buried) which determines the burial class.
The user can specify the values for the fraction correct cutoffs (the values below which absolute DDG values are considered neutral).

Adding a function to update the PDB JSON file (does not really belong in this file).