Note: Changing the author order and taking the Iron Throne by force.
This commit breaks everything. Intentionally. Rudely, even.

To prepare for cross-analyzing benchmark runs and to allow the published data from Kellogg et al. (where we only have the summary data) to be treated on par with generated data (where we have structures etc.), I am classing up the joint. Data from benchmark runs are now stored in objects. The plan is to load all of the data in a pandas dataframe rather than the traditional dicts we currently use. This should allow me to splice and scale the data easily and to generate the correlation, MAE, and fraction correct metrics over subsets of the data.

The new approach creates all analysis plots in one directory and leaves the input analysis data in the benchmark run directories. Each benchmark run used as input must have a unique, corresponding benchmark name which is then used to prefix generated files. This makes it easier for me to generate data for distribution.

Once I fix the code to work with the new format, I will:
 - compute the adjustment scalar as I do currently but promote it to be a main attribute of a run;
 - add an assertion that the analysis directory is not one of the benchmark directories (this will simplify reasoning);
 - add an option to copy the stored data (analysis_input.csv/.json, benchmark_data) into the analysis directory, prefixed with the associated benchmark name (this removes some manual copying/renaming when distributing data);
 - add a dataset section to the analysis file:
   - add data to the dataset JSON files which will be pulled into the analysis e.g. the name of the dataset, reference to the literature;
   - add pie-charts describing the properties of the dataset e.g. the percentage of mutations to alanine, the distribution of SCOP classes;
 - I will add comparative graphs between different benchmark runs;
 - I will add new slides with tabular data for the breakdown graphs e.g. for the buried/exposed graph"
  - The table will have three rows (B, E, N/A) and three sections of columns:
   - section 1 contains the a column for n = the number of cases;
   - section 2 contains three columns: R, MAE, Fraction correct);
   - section 3 contains two columns: MAE_adj, Fraction correct_adj where these values are calculated from the run data adjusted with a scalar to maximize fraction correct.
  - If a row represents a certain percentage of the dataset, the n column will be colored from yellow to dark red to indicate how reliable the results are. This may help to better identify pathologies e.g. if we do poorly overall on buried mutations but there are very few of these then that does not tell us much.